fawkes-robotino: image to run fawkes-robotino in RCLL simulation
This docker image is based on fedora-ros:23-jade and adds the Fawkes
domain release for the RoboCup Logistics League (RCLL), and the refbox
and simulation models, plugins, and worlds for the RCLL.

The image can be run with just
  docker run -it -v /tmp/.X11-unix:/tmp/.X11-unix \
    -e DISPLAY=unix$DISPLAY --name rcll-sim --privileged \
    timn/fawkes-robotino:2015-f23-jade

It will bring up the full simulation. Confer
https://trac.fawkesrobotics.org/wiki/Carologistics/Gazsim-Setup-2015#RunningaGame
on how to actually run the game from there.