gzweb: support replacing the websocket port
In our setup with a Kubernetes nginx ingress we want to have the
websocket accessed via a separate path, not as a separate port (even
though internally it actually is on a separate path).

Therefore, allow to set an environment variable that can be used to
arbitrarly substitute the port number. It does not change how the
container works, it just allows to adapt it to the outside facing
interface of the ingress by building the proper websocket URL.