fedora-ros: increase ROS master timeout
Several issues on the cluster were caused by a timeout between
instantiating the ROS master and then accessing it. Unfortunately, this
timeout is hardcoded into roslaunch and set to 10 seconds. On a loaded
cluster, this can be insufficient. Increase this to 120 seconds by
modifying the installed Python file via sed.

This is a workaround, but this would need to be fixed in roslaunch to be
parametrizable.